Update index.html
Issue(s): The short, short version.  I lost my free domains because Meta threatened Freenom (the last free provider of TLDs) that I used along with Cloudflare to lower traffic to GitHub, decrease loading times and traffic for you and your ISP, and provide better security.  

The short version: Meta / Facebook sued FreeNom and several other providers of free TLDs (Top Level Domains).  They said it's because the majority of activity coming from free TLDs is malicious, fraudulent, spamming, phishing, and often criminal.

Meta's complaint is a classic case of cherry-picking facts.  Even with all their money and power, the complaint leans on "evidence" that with any amount of scrutiny or logic falls apart almost completely.  Meta says that amongst TLDs with the most Phishing reported, 5 TLDs of the top 10 TLDs are FreeNom TLDs.

According to their statistics, the TLD with the most reported phishing is not only run by someone other than FreeNom, but it's also the most common and sought-after TLD ever.  The founder of Mega.com changed his last name to it. 

It's so synonymous with the internet that there was a financial bubble and crash named after it, the dot com. Most years, about half of all web traffic goes to .com domains. There are approximately 159 million .com TLDs registered currently.

Here is a screenshot of the table Meta provided:  https://raw.githubusercontent.com/brad457/images/main/Metas_uncited_top_10_TLDs_invovled_in_phishing_Freenom.png

It’s a list of the top 10 TLDs by phishing reported.  It includes a Rank of 1-10, the TLD name, what company operates the domain registry, the number of domains registered in each TLD, and the number of phishing domains reported in each TLD.

I also generated 2 images of pie/bar graphs. The first one provides a visual representation of the number of domains registered in each TLD among the top 10 list:
https://raw.githubusercontent.com/brad457/images/main/Meta_phishing_top_10_TLDs.png

The second one provides a visual representation of the number of domains registered in each of the TLDs that Freenom runs that are named in the complaint:
https://raw.githubusercontent.com/brad457/images/main/Meta_phishing_Freenom_TLDs.png

I created one last image that pulls this all together. 
https://raw.githubusercontent.com/brad457/images/main/meta_phishing_freenom_TLDs_with_percents_and_comments.png
I added just 1 column to the table of data they provided that calculates the percentage of domains with Phishing reported out of all domains in each TLD, and it provides some better context showing (at least what I believe) that the percentage of domains with Phishing reported for Freenom's TLDs is in line with most other domains. Any slight increase should be taken with a grain of salt as that increase includes lowering a cost barrier for accessing TLDs. The positive benefits of that are not some statistically insignificant amount that a company is trying to use for whatever justification they have landed on that their actions will increase profits.

Meta couldn't win this case, but they have the power and money and time to choke and kill Freenom.  So they won because they are Meta.